fix: refactoring, icon svgs have been converted to 'outline stroke' - WIP (still waiting for one)
