+++ src/common/svg/IconLocation.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { makeStyles } from 'tss-react/mui'
-import theme from '../styles/createTheme'
+import theme from '../../styles/createTheme'
 
 const IconLocation = () => {
   const { classes } = useStyles()
