+++ src/common/svg/IconNearMe.tsx
@@ -1,0 +1,48 @@
+import React from 'react'
+import { makeStyles } from 'tss-react/mui'
+import theme from '../../styles/createTheme'
+
+type IconNearMeProps = {
+  color?: string
+}
+const IconNearMe: React.FC<IconNearMeProps> = ({ color }) => {
+  const { classes } = useStyles()
+  color ??= theme.palette.primary.main
+
+  return (
+    <svg
+      preserveAspectRatio="none"
+      viewBox="0 0 31 37"
+      fill="none"
+      className={classes.iconNearMe}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M15.8369 2.00905C12.3804 2.00905 9.06546 3.41134 6.62134 5.90744C4.17722 8.40355 2.80413 11.789 2.80413 15.319C2.80413 20.482 6.0778 25.4062 9.54277 29.1414C11.2551 30.9874 12.9708 32.4966 14.2596 33.5446C14.9031 34.068 15.4382 34.4747 15.8105 34.7493C15.8194 34.7559 15.8282 34.7624 15.8369 34.7688C15.8456 34.7624 15.8544 34.7559 15.8633 34.7493C16.2356 34.4747 16.7707 34.068 17.4143 33.5446C18.703 32.4966 20.4187 30.9874 22.1311 29.1414C25.596 25.4062 28.8697 20.482 28.8697 15.319C28.8697 11.789 27.4966 8.40355 25.0525 5.90744C22.6084 3.41134 19.2934 2.00905 15.8369 2.00905ZM15.8369 35.9955C15.2913 36.8313 15.291 36.8311 15.2907 36.8309L15.2898 36.8303L15.287 36.8284L15.2777 36.8219L15.2441 36.7987C15.2151 36.7787 15.1733 36.7495 15.1195 36.7114C15.0118 36.6352 14.8559 36.5235 14.6588 36.378C14.2645 36.0873 13.7046 35.6615 13.0341 35.1162C11.6948 34.027 9.90639 32.4546 8.11467 30.5232C4.57143 26.7035 0.836914 21.2895 0.836914 15.319C0.836914 11.2562 2.41727 7.3597 5.23031 4.48683C8.04336 1.61396 11.8587 0 15.8369 0C19.8152 0 23.6305 1.61396 26.4435 4.48683C29.2566 7.3597 30.8369 11.2562 30.8369 15.319C30.8369 21.2895 27.1024 26.7035 23.5592 30.5232C21.7674 32.4546 19.979 34.027 18.6397 35.1162C17.9692 35.6615 17.4093 36.0873 17.0151 36.378C16.8179 36.5235 16.662 36.6352 16.5544 36.7114C16.5005 36.7495 16.4587 36.7787 16.4298 36.7987L16.3962 36.8219L16.3868 36.8284L16.384 36.8303L16.3831 36.8309C16.3828 36.8311 16.3825 36.8313 15.8369 35.9955ZM15.8369 35.9955L16.3831 36.8309C16.0527 37.0558 15.6211 37.0558 15.2907 36.8309L15.8369 35.9955Z"
+        fill={color}
+        className={classes.iconTransition}
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M15.8369 11.552C13.7998 11.552 12.1484 13.2386 12.1484 15.319C12.1484 17.3994 13.7998 19.086 15.8369 19.086C17.874 19.086 19.5254 17.3994 19.5254 15.319C19.5254 13.2386 17.874 11.552 15.8369 11.552ZM10.1812 15.319C10.1812 12.129 12.7133 9.54299 15.8369 9.54299C18.9605 9.54299 21.4927 12.129 21.4927 15.319C21.4927 18.509 18.9605 21.095 15.8369 21.095C12.7133 21.095 10.1812 18.509 10.1812 15.319Z"
+        fill={color}
+        className={classes.iconTransition}
+      />
+      className={classes.iconTransition}
+    </svg>
+  )
+}
+export default IconNearMe
+const useStyles = makeStyles()(() => {
+  return {
+    iconNearMe: {
+      maxWidth: '100%',
+      maxHeight: '100%',
+    },
+    iconTransition: {
+      transition: 'color .3s, stroke .3s, fill .3s',
+    },
+  }
+})
