+++ src/helpers/navigationConfigHelper.ts
@@ -1,9 +1,9 @@
 import { NavigationConfig } from '../common/types/navigationTypes'
-import IconChat from '../common/IconChat'
-import iconLightning from '../common/IconLightning'
-import iconII from '../common/IconII'
-import IconProfile from '../common/IconProfile'
-import IconNearMe from '../common/IconNearMe'
+import IconChat from '../common/svg/IconChat'
+import iconLightning from '../common/svg/IconLightning'
+import iconII from '../common/svg/IconII'
+import IconProfile from '../common/svg/IconProfile'
+import IconNearMe from '../common/svg/IconNearMe'
 
 export const generateNavigationConfig = (): NavigationConfig[] => {
   // can be optimized by using only one config, common for footer and header (needed: changes in canvas for icons, making the images of one size, hiding user's icon on desktop - by ?class )
@@ -13,13 +13,13 @@ export const generateNavigationConfig = (): NavigationConfig[] => {
       value: 'nearme',
       Icon: IconNearMe,
       linkTo: '/user/near-me',
-      size: { xs: { width: 18, height: 22 }, lg: { width: 30, height: 35 } },
+      size: { xs: { width: 18, height: 22 }, lg: { width: 30, height: 37 } },
     },
     {
       value: 'wholikedyou',
       Icon: iconLightning,
       linkTo: '/user/who-liked-you',
-      size: { xs: { width: 18, height: 24 }, lg: { width: 28, height: 38 } },
+      size: { xs: { width: 18, height: 24 }, lg: { width: 29, height: 40 } },
     },
     {
       value: 'friends',
